refactor: move Zacks provider to provider/zacks sub-package
