+++ cmd/providers_register.go
@@ -21,4 +21,5 @@ import (
 	_ "github.com/penny-vault/pvdata/provider/massive"
 	_ "github.com/penny-vault/pvdata/provider/nasdaq"
 	_ "github.com/penny-vault/pvdata/provider/tiingo"
+	_ "github.com/penny-vault/pvdata/provider/zacks"
 )
